+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Implemented OCLC other editions call
 - Added functionality for locally generating Counter 5 downloads reports to analytics folder
 - Readded enter and exit functions to API DB client
+- Refactored works and editions APIs and added error handling
 
 ## 2024-08-06 -- v0.13.1
 ## Added